+++ src/scopes/internal/RegistryObject.cpp
@@ -20,6 +20,7 @@
 
 #include <unity/scopes/internal/MWRegistry.h>
 #include <unity/scopes/internal/RuntimeImpl.h>
+#include <unity/scopes/internal/Utils.h>
 #include <unity/scopes/ScopeExceptions.h>
 #include <unity/UnityExceptions.h>
 #include <unity/util/ResourcePtr.h>
@@ -34,6 +35,9 @@
 #include <wordexp.h>
 
 using namespace std;
+
+static const char* c_debug_dbus_started_cmd = "dbus-send --type=method_call --dest=com.ubuntu.SDKAppLaunch /ScopeRegistryCallback com.ubuntu.SDKAppLaunch.ScopeLoaded";
+static const char* c_debug_dbus_stopped_cmd = "dbus-send --type=method_call --dest=com.ubuntu.SDKAppLaunch /ScopeRegistryCallback com.ubuntu.SDKAppLaunch.ScopeStopped";
 
 namespace unity
 {
@@ -340,6 +344,7 @@ void RegistryObject::on_state_received(std::string const& scope_id, StateReceive
 RegistryObject::ScopeProcess::ScopeProcess(ScopeExecData exec_data, MWPublisher::SPtr publisher)
     : exec_data_(exec_data)
     , reg_publisher_(publisher)
+    , manually_started_(false)
 {
 }
 
@@ -452,8 +457,8 @@ void RegistryObject::ScopeProcess::exec(
         env["LD_LIBRARY_PATH"] = scope_ld_lib_path;  // Overwrite any LD_LIBRARY_PATH entry that may already be there.
 
         process_ = executor->exec(program, argv, env,
-                                     core::posix::StandardStream::stdin | core::posix::StandardStream::stdout,
-                                     exec_data_.confinement_profile);
+                                  core::posix::StandardStream::stdin | core::posix::StandardStream::stdout,
+                                  exec_data_.confinement_profile);
         if (process_.pid() <= 0)
         {
             clear_handle_unlocked();
@@ -523,28 +528,19 @@ void RegistryObject::ScopeProcess::update_state_unlocked(ProcessState new_state)
     }
     else if (new_state == Running)
     {
-        if (reg_publisher_)
-        {
-            // Send a "started" message to subscribers to inform them that this scope (topic) has started
-            reg_publisher_->send_message("started", exec_data_.scope_id);
-        }
+        publish_state_change(Running);
 
         if (state_ != Starting)
         {
             cout << "RegistryObject::ScopeProcess: Process for scope: \"" << exec_data_.scope_id
                  << "\" started manually" << endl;
 
-            // Don't update state_, treat this scope as not running if a locate() is requested
-            return;
+            manually_started_ = true;
         }
     }
     else if (new_state == Stopped)
     {
-        if (reg_publisher_)
-        {
-            // Send a "stopped" message to subscribers to inform them that this scope (topic) has stopped
-            reg_publisher_->send_message("stopped", exec_data_.scope_id);
-        }
+        publish_state_change(Stopped);
 
         if (state_ != Stopping)
         {
@@ -552,19 +548,15 @@ void RegistryObject::ScopeProcess::update_state_unlocked(ProcessState new_state)
                  << "\" closed unexpectedly. Either the process crashed or was killed forcefully." << endl;
         }
     }
-    else if (new_state == Stopping && state_ != Running)
+    else if (new_state == Stopping && manually_started_)
     {
-        if (reg_publisher_)
-        {
-            // Send a "stopped" message to subscribers to inform them that this scope (topic) has stopped
-            reg_publisher_->send_message("stopped", exec_data_.scope_id);
-        }
+        publish_state_change(Stopped);
 
         cout << "RegistryObject::ScopeProcess: Manually started process for scope: \""
              << exec_data_.scope_id << "\" exited" << endl;
 
-        // Don't update state_, treat this scope as not running if a locate() is requested
-        return;
+        new_state = Stopped;
+        manually_started_ = false;
     }
     state_ = new_state;
     state_change_cond_.notify_all();
@@ -662,6 +654,50 @@ std::vector<std::string> RegistryObject::ScopeProcess::expand_custom_exec()
     }
 
     return command_args;
+}
+
+void RegistryObject::ScopeProcess::publish_state_change(ProcessState scope_state)
+{
+    if (scope_state == Running)
+    {
+        if (reg_publisher_)
+        {
+            // Send a "started" message to subscribers to inform them that this scope (topic) has started
+            reg_publisher_->send_message("started", exec_data_.scope_id);
+        }
+        if (exec_data_.debug_mode)
+        {
+            // If we're in debug mode, callback to the SDK via dbus (used to monitor scope lifecycle)
+            std::string started_message = c_debug_dbus_started_cmd;
+            started_message += " string:" + exec_data_.scope_id + " uint64:" + std::to_string(process_.pid());
+            if (safe_system_call(started_message) != 0)
+            {
+                std::cerr << "RegistryObject::ScopeProcess::publish_state_change(): "
+                             "Failed to execute SDK DBus ScopeLoaded callback "
+                             "(Scope ID: " << exec_data_.scope_id << ")" << endl;
+            }
+        }
+    }
+    else if (scope_state == Stopped)
+    {
+        if (reg_publisher_)
+        {
+            // Send a "stopped" message to subscribers to inform them that this scope (topic) has stopped
+            reg_publisher_->send_message("stopped", exec_data_.scope_id);
+        }
+        if (exec_data_.debug_mode)
+        {
+            // If we're in debug mode, callback to the SDK via dbus (used to monitor scope lifecycle)
+            std::string stopped_message = c_debug_dbus_stopped_cmd;
+            stopped_message += " string:" + exec_data_.scope_id;
+            if (safe_system_call(stopped_message) != 0)
+            {
+                std::cerr << "RegistryObject::ScopeProcess::publish_state_change(): "
+                             "Failed to execute SDK DBus ScopeStopped callback "
+                             "(Scope ID: " << exec_data_.scope_id << ")" << endl;
+            }
+        }
+    }
 }
 
 } // namespace internal