+++ src/scopes/internal/smartscopes/SSRegistryObject.cpp
@@ -25,6 +25,7 @@
 #include <unity/scopes/internal/ScopeImpl.h>
 #include <unity/scopes/internal/ScopeMetadataImpl.h>
 #include <unity/scopes/internal/smartscopes/HttpClientQt.h>
+#include <unity/scopes/internal/Utils.h>
 #include <unity/scopes/ScopeExceptions.h>
 #include <unity/UnityExceptions.h>
 
@@ -87,7 +88,7 @@ SSRegistryObject::SSRegistryObject(MiddlewareBase::SPtr middleware,
     refresh_thread_ = std::thread(&SSRegistryObject::refresh_thread, this);
 }
 
-SSRegistryObject::~SSRegistryObject() noexcept
+SSRegistryObject::~SSRegistryObject()
 {
     // stop the refresh thread
     {
@@ -307,7 +308,7 @@ void SSRegistryObject::get_remote_scopes()
                     {
                         // Store both JSON (for internal comparison) and DB (for external use)
                         changed = true;
-                        std::string settings_db = RuntimeConfig::default_config_directory() + "/" + scope.id + "/settings.db";
+                        std::string settings_db = RuntimeConfig::default_config_directory() + "/" + scope.id + "/settings.ini";
                         SettingsDB::SPtr db(SettingsDB::create_from_schema(settings_db, *schema));
                         settings_defs_[scope.id] = SSSettingsDef{settings, db, needs_location_data};
                     }
@@ -374,7 +375,7 @@ void SSRegistryObject::get_remote_scopes()
     if (changed)
     {
         // something has changed, send invalidate signal
-        int result = system(c_dbussend_cmd);
+        int result = safe_system_call(c_dbussend_cmd);
         (void)result;
     }
 }