+++ RELEASE_NOTES.md
@@ -1,11 +1,22 @@
 Release notes
 =============
 
+Changes in version 0.6.5
+========================
+  - Implemented support for expandable preview widgets. See the documentation of PreviewWidget
+    for details of the new widget type.
+
+Changes in version 0.6.3
+========================
+  - Fix scope cache path for confined scopes.
+
 Changes in version 0.6.2
 ========================
   - Move scope configuration to ~/.config/unity-scopes/
 
   - New setting to enable/disable of location data being fed to scopes.
+
+  - New DebugMode scope configuration option.
  
 Changes in version 0.6.1
 ========================